feat: add share reason in share requests (#498)
### Feature or Bugfix
- Feature


### Detail
- Expose the Share Request Comments as Part of the Share View for the
Approver to See
- Allow Approver to Provide Comments when Rejecting a Share for the
Requester to See

### Relates
- [#422 ]

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

+++ backend/dataall/db/api/dataset.py
@@ -358,6 +358,26 @@ class Dataset:
 
     @staticmethod
     def transfer_stewardship_to_owners(session, dataset):
+        # Remove Steward Resource Policy on Dataset
+        env = Environment.get_environment_by_uri(session, dataset.environmentUri)
+        if dataset.stewards != env.SamlGroupName:
+            ResourcePolicy.delete_resource_policy(
+                session=session,
+                group=dataset.stewards,
+                resource_uri=dataset.datasetUri,
+            )
+
+        # Remove Steward Resource Policy on Dataset Tables
+        dataset_tables = [t.tableUri for t in Dataset.get_dataset_tables(session, dataset.datasetUri)]
+        for tableUri in dataset_tables:
+            if dataset.stewards != env.SamlGroupName:
+                ResourcePolicy.delete_resource_policy(
+                    session=session,
+                    group=dataset.stewards,
+                    resource_uri=tableUri,
+                )
+
+        # Remove Steward Resource Policy on Dataset Share Objects
         dataset_shares = (
             session.query(models.ShareObject)
             .filter(models.ShareObject.datasetUri == dataset.datasetUri)
@@ -365,19 +385,17 @@ class Dataset:
         )
         if dataset_shares:
             for share in dataset_shares:
-                ResourcePolicy.attach_resource_policy(
+                ResourcePolicy.delete_resource_policy(
                     session=session,
-                    group=dataset.SamlAdminGroupName,
-                    permissions=permissions.SHARE_OBJECT_APPROVER,
+                    group=dataset.stewards,
                     resource_uri=share.shareUri,
-                    resource_type=models.ShareObject.__name__,
                 )
         return dataset
 
     @staticmethod
     def transfer_stewardship_to_new_stewards(session, dataset, new_stewards):
         env = Environment.get_environment_by_uri(session, dataset.environmentUri)
-        if dataset.stewards != env.SamlGroupName:
+        if dataset.stewards != dataset.SamlAdminGroupName:
             ResourcePolicy.delete_resource_policy(
                 session=session,
                 group=dataset.stewards,
@@ -393,7 +411,7 @@ class Dataset:
 
         dataset_tables = [t.tableUri for t in Dataset.get_dataset_tables(session, dataset.datasetUri)]
         for tableUri in dataset_tables:
-            if dataset.stewards != env.SamlGroupName:
+            if dataset.stewards != dataset.SamlAdminGroupName:
                 ResourcePolicy.delete_resource_policy(
                     session=session,
                     group=dataset.stewards,
@@ -421,11 +439,12 @@ class Dataset:
                     resource_uri=share.shareUri,
                     resource_type=models.ShareObject.__name__,
                 )
-                ResourcePolicy.delete_resource_policy(
-                    session=session,
-                    group=dataset.stewards,
-                    resource_uri=share.shareUri,
-                )
+                if dataset.stewards != dataset.SamlAdminGroupName:
+                    ResourcePolicy.delete_resource_policy(
+                        session=session,
+                        group=dataset.stewards,
+                        resource_uri=share.shareUri,
+                    )
         return dataset
 
     @staticmethod
